Class restructuring + search improvement
The code has been restructured such that it includes a "Site" class,
each instance of which holds the variables that were previously stored
in arrays in Container.  Container has an array of Sites, called
"sites".

The search functionality has been improved both organizationally in the
way that files are downloaded as well as to support things like multiple
pages of results.