Fixes and efficiency improvements
Due to an oversight on my part, the code was not in working order during
the last commit, so I fixed that.  Also, the program now stops looking
through the HTML of a page as soon as it finds what it's looking for,
instead of running to the end of the page, which should bring some drop
in the amount of time required for the program to run, especially on
pages with many lines of HTML.

One of the main bottlenecks in the program's speed (other than file
downloads, obviously), seems to be the loading of the web pages.